requirements.txt
python-dotenv==1.0.1
requests==2.32.3
flask==3.0.3
redis==5.2.0
Flask-Limiter==3.8.0